+++ index.js
@@ -3,7 +3,6 @@ const inquirer = require("inquirer");
 const fs = require("fs");
 const shapes = require("./lib/shapes.js");
 const shapesTest = require("./lib/shapes.test.js");
-const createLogo = require("./lib/shapes");
 
 inquirer
     .prompt([
@@ -43,20 +42,7 @@ inquirer
         },
     ])
 
-    //Use fs towrite the file to logo.svg.  
-    //Wrap response in shapes() function passing in responses.  This has been required in at top.
-    //throw error if not generated, also console.log message to user about colors 
-    //otherwise console.log successful generation if it is generated
-    //I'M NOT CALLING THE FUNCTION I WANT CORRECTLY
-    .then((response) => {
-        //create new variable with response wrapped in createLogo function from other js page
-        const logoData = createLogo(response);
-        fs.writeFile("logo.svg", logoData,
-            (err) => err ? (console.error(err),
-                console.log("There was an error in generating your file.  Please check your color spelling, visit 147colors.com to reference available colors. Please try again!"))
-
-                : console.log("Generated logo.svg"))
-    })
+    //How do these responses get passed and where do they need to get passed?  Just to createLogo?  Test that first.
 
 
 /*
