+++ lib/createLogo.js
@@ -1,0 +1,14 @@
+const Triangle = require('./utils/triangle.js')
+const Square = require('./utils/square.js')
+const Circle = require('./utils/circle.js')
+const fs = require('fs');
+
+// .then((response) => {
+//     //create new variable with response wrapped in createLogo function from other js page
+//     const logoData = createLogo(response);
+//     fs.writeFile("logo.svg", logoData,
+//         (err) => err ? (console.error(err),
+//             console.log("There was an error in generating your file.  Please check your color spelling, visit 147colors.com to reference available colors. Please try again!"))
+
+//             : console.log("Generated logo.svg"))
+// })
