+++ lib/shapes.js
@@ -1,54 +1,21 @@
 class Shape {
     //Parent class
-    constructor(logoLetters, logoLettersColor, logoShape, logoColor) {
+    constructor(logoLetters, logoLettersColor, logoShape, logoColor) 
+    {
         this.logoLetters = logoLetters;
         this.logoLettersColor = logoLettersColor;
         this.logoShape = logoShape;
         this.logoColor = logoColor;
-        // Added this for the y axis plot of the text.  Setting to 125 for both square and circle, triangle has it's own value to overwrite
-        this.yPlot = 125;
-    }
-   
+    }   
     }
 
 
 //All three individual classes EXTEND the parent class of Shape
-class Circle extends Shape {
-    // Using constructor to assign the object properties. 
-    //DO I NEED TO PULL EVERYTHING OR JUST logoShape?
-    constructor(logoLetters, logoLettersColor, logoShape, logoColor) {
-        // Using super to inherit from parent
-        super(logoLetters, logoLettersColor, logoShape, logoColor);
-        //this is the shape input of a circle.  Keep space at end to join into string with fill...
-        this.shapeSection = '<circle cx="150" cy="100" r="80" '
-    }
-    render() {
-       //combined shapeSection and logoColor to bundle whole line.  Cleaner in my head.
-        return `${this.shapeSection} fill="${this.logoColor}" />`;
-    }
-}
 
-class Square extends Shape {
-    constructor(logoLetters, logoLettersColor, logoShape, logoColor) {
-        super(logoLetters, logoLettersColor, logoShape, logoColor);
-        //replaced with shape input of Square (rectangle with same width and height)
-        this.shapeSection = '<rect x="75" y="25" width="150" height="150" '
-    }
-    render() {
-        return `${this.shapeSection} fill="${this.logoColor}" />`;
-}}
 
-class Triangle extends Shape {
-    constructor(logoLetters, logoLettersColor, logoShape, logoColor) {
-        super(logoLetters, logoLettersColor, logoShape, logoColor);
-        //replaced with shape input of Triangle
-        this.shapeSection = '<polygon points="150, 18 244, 182 56, 182" '
-        //extra step on this shape.  Change yPlot to 150
-        this.yPlot = 150;
-    }
-    render() {
-        return `${this.shapeSection} fill="${this.logoColor}" />`;
-}}
+
+
+
 
 //NOW TO ADD THE createShape function
 function createLogo(response) {
